azure_function_ppt_v2/requirements.txt
# Azure Functions Core Dependencies
azure-functions>=1.18.0
azure-core>=1.25.0

# Semantic Kernel for AI Agents
semantic-kernel==1.30.0

# Azure OpenAI Integration
openai>=1.0.0

# Configuration and Environment
python-dotenv>=1.0.0

# JSON and Data Processing
json5>=0.9.10

# Date and Time Utilities
python-dateutil>=2.8.2

# HTTP Requests (if needed for external APIs)
requests>=2.28.0

# File Processing Utilities
pathlib2>=2.3.7

# Base64 Encoding/Decoding
base64io>=1.0.3

# UUID Generation
uuid>=1.30

# Regular Expressions (built-in, but explicit for clarity)
# re (built-in)

# Subprocess Management (built-in, but critical for Pandoc)
# subprocess (built-in)

# Temporary File Handling (built-in, but critical for file operations)  
# tempfile (built-in)

# Operating System Interface (built-in, but needed for file operations)
# os (built-in)

# IMPORTANT: Pandoc must be installed separately on the system
# For Azure Functions, this would need to be included in the deployment
# Consider using a custom Docker container or build script

# Development and Testing Dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Type Hints Support
typing-extensions>=4.4.0